+++ TeamProject/src/main/java/com/restinginbed/TeamProject/jpa_repositories/ItemRepository.java
@@ -1,6 +1,7 @@
 package com.restinginbed.TeamProject.jpa_repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.restinginbed.TeamProject.Item;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 }
